add errors checking & logs in service

+++ src/main/java/com/sorb/mybatisapp/controller/RegionController.java
@@ -2,7 +2,6 @@ package com.sorb.mybatisapp.controller;
 
 import com.sorb.mybatisapp.model.Region;
 import com.sorb.mybatisapp.service.RegionService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -28,24 +27,47 @@ public class RegionController {
 
     @GetMapping("/{id}")
     public ResponseEntity<Region> getRegion(@PathVariable int id) {
+        if (!isExists(id)) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         return new ResponseEntity<>(regionService.getRegionById(id), HttpStatus.OK);
     }
 
     @PostMapping
     public ResponseEntity<Region> createRegion(@RequestBody Region region) {
+
+        if (!isRegionFilled(region)) return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         Integer res = regionService.createRegion(region);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
 
     @PatchMapping("/{id}")
     public ResponseEntity<Region> updateRegion(@PathVariable int id, @RequestBody Region region) {
+        if (!isExists(id)) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         regionService.updateRegion(region, id);
         return new ResponseEntity<>(HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Region> deleteRegion(@PathVariable int id){
+    public ResponseEntity<Region> deleteRegion(@PathVariable int id) {
+
+        if (!isExists(id)) return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
         regionService.deleteRegion(id);
         return new ResponseEntity<>(HttpStatus.OK);
+    }
+
+    private boolean isExists(int id) {
+        Region region = regionService.getRegionById(id);
+        return region != null;
+    }
+
+    private boolean isRegionFilled(Region region) {
+        boolean result;
+        try {
+            result = region.getFullName() != null && region.getShortName() != null
+                    && !region.getShortName().trim().isEmpty() && !region.getFullName().trim().isEmpty();
+        } catch (NullPointerException e) {
+            return false;
+        }
+        return result;
     }
 }
